+++ endstone_bot/__init__.py
@@ -32,7 +32,7 @@ from __future__ import annotations
 from endstone_bot.bot_plugin import BotPlugin
 
 __all__ = ["BotPlugin"]
-__version__ = "3.2.2"
+__version__ = "3.2.3"
 
 REFERENCES = {
     "mcbes_manage_script": {